DPDK: Use generic CPU instructions and skip AVX512
This commits removes AVX512 support and enables generic CPU options for
the DPDK build. The IPDK CI builds on Azure VMs which support AVX512.
However, since Virtualbox doesn't support this (see [1] and [2]), we
can skip this for p4-dpdk-target when building DPDK as a part of the
p4-dpdk-target repo.

Fixes: https://github.com/ipdk-io/ipdk/issues/118

[1] https://forums.virtualbox.org/viewtopic.php?t=86047
[2] https://forums.virtualbox.org/viewtopic.php?f=7&t=87131

Signed-off-by: Kyle Mestery <mestery@mestery.com>