NVFMSA-4808 | connection flow | support to retrieve entries from HW.
This commit is to support iterating from HW table directly in TDI
(this support already exist with BFRT).

Port following commit ab3fe66c0211da70ebe88ed3bc7c654c86b4dfd7

Implemented read iterator for P4 table to directly read from HW.
Entries are not stored in the SDE for table with add_on_miss enabled
or if the entries size is greater than a threshold (10000 entries)
for LEM and SEM block.
Only MatchKeys are returned for the iterator APIs when reading
from HW directly. Actions are currently shown as NoAction and will
be updated in later commit.

Signed-off-by: riyazahm22 <riyaz.ahmad@intel.com>