Address rule encoding failure issue
Encoding rule is failing when the order of match key fields in
the table are differ than what they appear in header.

Issue is because while calculating offset value, considering first match
key field position as base offset irrespective of where that field exist
in the header, which leads to rule encoding failure.

To mitigate the issue, calculating base offset value properly.

Signed-off-by: Anand Sunkad <anand.sunkad@intel.com>