Set up a basic automated e2e smoke test (#102)
* Add a Dockerfile for a reproducible environment

The final image is 5.3GB. That seems too large. There should be room for optimization.

* Add a simple testing pipeline.

The testing scripts are not finished yet. But I ran into some issues that might need to be resolved before I can move on.

* Update TDI and DPDK submodules.

TDI is updated to the latest commit (on Mar 1, 2023). DPDK is updated to the latest release (v23.03).

* Convert ad-hoc shell scripts into reusable Python scripts

* Add the bfshell command

* Set up a basic automated smoke test

There are still lots of issues to solve before this smoke test becomes meaningful enough. But this could be a good starting point.

* Switch back to original versions of DPDK and TDI in this repo

This brings back the enable() method that seems to be a crucial step in setting up the pipeline. But this also results in some errors. Will try to fix them.

* Make the test case pass

This is a temporary solution though.

* General refinement

* Add forwarding table back to the P4 pipeline

It works this time!

* Log command and returned status

* Use zlog-cfg file to configure log level instead of hard coding

* Switch to checking file existence as a temporary workaround

* Use multi-stage build to reduce Docker image size

Size is reduced from 4.78GB to 1GB! Thanks @chrispsommers for the suggestion.